Refactor: Replace SudokuParser with Enhanced SudokuFormatHandler
This commit involves a significant refactor of the Sudoku parsing functionality. The previous SudokuParser class is replaced with a more versatile and modular SudokuFormatHandler. This new handler introduces flexibility in handling different input types (file paths and strings) for parsing Sudoku boards. The key changes include:

- Introduction of the abstract base class FormatHandler with abstract methods parse and save.
- Implementation of specific format handlers like GridFormatHandler and FlatFormatHandler to accommodate different Sudoku board formats.
- Addition of input type validation and preprocessing steps to streamline parsing.
- Improved file handling and format validation logic.
- Moved the saving logic to the format class instead of the SudokuBoard class as in t  the context of multiple file formats, it makes sense to implement the saving logic   in the same area where the new file format is being supported. This way one can
  save boards in many formats.
- Added basic implementation of one line board format in LineFormatHandler class
- Put some common pre processing functions for file parsing in utils

Also added a minor change to SudokuBoards board validation process, just incase
someone instantiates the class directly with a rogue array, checks the array
contains integers between 0 and 9 along with the other validation checks.

This refactor enhances the extensibility of the Sudoku board parsing functionality, making it easier to support additional formats and input types in the future.