Optimise Sudoku Solver with Set-Based Validity Checks
This commit introduces a significant optimization to the Sudoku solving algorithm by employing sets for tracking the used numbers
in each row, column, and 3x3 subgrid. This change replaces the previous method of looping through each row, column, and subgrid
to check for the presence of a number, which was found as a significant bottleneck to the backtracking method  by lineprofiler.

Key changes include:
- Addition of three arrays of sets as class attributes in SudokuBoard: `rows`, `columns`, and `subgrids`.
- Modification of the `check_valid` method to utilize these sets.
- Implementation of `place_number` and `remove_number` methods in SudokuBoard to update these sets whenever a number is placed on
  or removed from the board.
- Adjustment in the backtracking algorithm (`_backtrack` method) to utilise the updated `check_valid` method.

This optimisation significantly reduces the computational complexity of the solver, particularly noticeable in the profiling results
where the runtime of the `check_valid` method is drastically decreased by almost 90%. This leads to an overall performance improvement
in solving Sudoku puzzles.