Change get_empty method of SudokuBoard to find_empty_cells and move save board test
Updated the get_empty method of the SudokuBoard to find_empty_cells to return a list that can be
looped over to check for cells which are easier to solve. Useful for when trying to implement constraints.
Updated the backtracking method in BacktrackingSolver and updated the unit tests.

Moved the save board unit test from the sudoku_board unit tests to the sudoku_format_handler unit tests
inline with the new class refactor.