Remove custom git hook from pre commit and set git hash in main.py
Git hash was not being set as an environment variable for the current shell when
set inside pre commit. Easier to set it during main.py call.

Changed file format for the summary file to a json to save the runtime arguments as well.

Added runtime args and git hash to stats file for single input run as well but file format
needs modification. Currently mixing txt file and json, not recommended long term.