Remove verbose cli arg from main.py
Realised that only 2 run modes, batch and single input.
For batch verbose mode is not supported as it slows down the script and will clog the terminal.
For single input it is always useful so redundant flag since it is always set and no usecase where
it would be detrimental to have on for single puzzle solving.