Change get_related_cell_values method to find_possible_cell_values for SudokuBoard class
Change method function from fetching the set of taken options, i.e options for values the cell can not take on
to the set of possible options for that cell. This is the more useful quantity to return as it is what you ultimately care about.

Also Renamed method for clarity: 'get_related_cell_values' to 'find_possible_cell_values' in SudokuBoard, enhancing readability,
aligning with Sudoku terminology, improving method discoverability, and providing semantic precision.