Enhanced Encapsulation: Removed SudokuBoard Setter, Updated Place Number Method and SudokuBoard unit tests.
Strengthened the encapsulation of the SudokuBoard class by removing the direct setter method. This change ensures that the internal state
of the board is not externally modified in an uncontrolled manneR. Furthermore, we have refined the place_number method
to include a validation check, ensuring that any number placed on the board adheres to the  rules of Sudoku.