Bugfix for invalid board detection and for single input mode
Fixed an issue where invalid boards with no solution were not being detected
because the original SudokuBoard object was not equal to the solved board object,
moved to checking their array attributes. Modified single input mode in src/main.py to call the
same function that is called to solve boards as in batch mode. Making the code more maintainable and modular.