Refactor main and move functions to utils part II
Forgot to add main to previous commit, this main has some of its previous
functions moved to utilities for a more clean modular design. The only function
left in main is the validate_args function which is intimately tied to the
the front end and so is left in main.