Bugfix for saving and batch processing in main.py using load_boards
Load_boards was not being called properly by main.py
Output save paths were also incorrectly being handled as the logic had not
been updated to be consistent with the output of load_boards being the
full path of the boards and not just the filename.

Tidied up docstring of load board for the output type to be signposted
better. Also updated error raised part of docstring for improved transparency of behaviour.
Moved single processing inside the else, rather than being set at the start
which was not clear behaviour.